+++ application/tests/UnitTestCase.php
@@ -1,0 +1,5 @@
+<?php
+
+class UnitTestCase extends CIPHPUnitTestUnitTestCase
+{
+}